+++ web/js/brand.js
@@ -1,0 +1,2 @@
+/* This file is just a stub to test the browserify build task */
+export default function () {};
