+++ web/js/index.js
@@ -1,1 +1,3 @@
 export { default as stub } from './stub';
+export { default as polyfill } from './polyfill';
+export { default as util } from './util/util';
